Linux crash (#9)
* Qbs build: correct rpath for conan libraries

Libraries managed with conan are now downloaded in dynamic. They are
also installed in the build directory.
This directory path is also added to all the applications rpaths.
This ensure that the application loads the correct library.

* Qbs build: correct rpath for conan libraries

Libraries managed with conan are now downloaded in dynamic. They are
also installed in the build directory.
This directory path is also added to all the applications rpaths.
This ensure that the application loads the correct library.